requirements.txt
Click==7.0
dataclasses==0.7
dnspython==1.16.0
email-validator==1.0.5
fastapi==0.42.0
h11==0.8.1
httptools==0.0.13
idna==2.8
pkg-resources==0.0.0
psycopg2==2.8.4
pydantic==0.32.2
python-dotenv==0.10.3
starlette==0.12.9
uvicorn==0.10.8
uvloop==0.14.0
v==0.0.0
websockets==8.1
